rbd: dissect the krbd-based backend from RBDUtil into RBDKernel.
