+++ pkg/volume/rbd/rbd_util.go
@@ -23,75 +23,21 @@ package rbd
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
-	"io/ioutil"
 	"math/rand"
 	"os"
 	"path"
-	"regexp"
 	"strings"
-	"time"
 
 	"github.com/golang/glog"
 	"k8s.io/kubernetes/pkg/api/v1"
-	"k8s.io/kubernetes/pkg/util/exec"
 	"k8s.io/kubernetes/pkg/util/mount"
-	"k8s.io/kubernetes/pkg/util/node"
 	"k8s.io/kubernetes/pkg/volume"
 )
 
 const (
 	imageWatcherStr = "watcher="
-	kubeLockMagic   = "kubelet_lock_magic_"
 )
-
-// search /sys/bus for rbd device that matches given pool and image
-func getDevFromImageAndPool(pool, image string) (string, bool) {
-	// /sys/bus/rbd/devices/X/name and /sys/bus/rbd/devices/X/pool
-	sys_path := "/sys/bus/rbd/devices"
-	if dirs, err := ioutil.ReadDir(sys_path); err == nil {
-		for _, f := range dirs {
-			// pool and name format:
-			// see rbd_pool_show() and rbd_name_show() at
-			// https://github.com/torvalds/linux/blob/master/drivers/block/rbd.c
-			name := f.Name()
-			// first match pool, then match name
-			po := path.Join(sys_path, name, "pool")
-			img := path.Join(sys_path, name, "name")
-			exe := exec.New()
-			out, err := exe.Command("cat", po, img).CombinedOutput()
-			if err != nil {
-				continue
-			}
-			matched, err := regexp.MatchString("^"+pool+"\n"+image+"\n$", string(out))
-			if err != nil || !matched {
-				continue
-			}
-			// found a match, check if device exists
-			devicePath := "/dev/rbd" + name
-			if _, err := os.Lstat(devicePath); err == nil {
-				return devicePath, true
-			}
-		}
-	}
-	return "", false
-}
-
-// stat a path, if not exists, retry maxRetries times
-func waitForPath(pool, image string, maxRetries int) (string, bool) {
-	for i := 0; i < maxRetries; i++ {
-		devicePath, found := getDevFromImageAndPool(pool, image)
-		if found {
-			return devicePath, true
-		}
-		if i == maxRetries-1 {
-			break
-		}
-		time.Sleep(time.Second)
-	}
-	return "", false
-}
 
 // make a directory like /var/lib/kubelet/plugins/kubernetes.io/pod/rbd/pool-image-image
 func makePDNameInternal(host volume.VolumeHost, pool string, image string) string {
@@ -102,94 +48,6 @@ type RBDUtil struct{}
 
 func (util *RBDUtil) MakeGlobalPDName(rbd rbd) string {
 	return makePDNameInternal(rbd.plugin.host, rbd.Pool, rbd.Image)
-}
-
-func (util *RBDUtil) rbdLock(b rbdMounter, lock bool) error {
-	var err error
-	var output, locker string
-	var cmd []byte
-	var secret_opt []string
-
-	if b.Secret != "" {
-		secret_opt = []string{"--key=" + b.Secret}
-	} else {
-		secret_opt = []string{"-k", b.Keyring}
-	}
-	if len(b.adminId) == 0 {
-		b.adminId = b.Id
-	}
-	if len(b.adminSecret) == 0 {
-		b.adminSecret = b.Secret
-	}
-
-	// construct lock id using host name and a magic prefix
-	lock_id := kubeLockMagic + node.GetHostname("")
-
-	l := len(b.Mon)
-	// avoid mount storm, pick a host randomly
-	start := rand.Int() % l
-	// iterate all hosts until mount succeeds.
-	for i := start; i < start+l; i++ {
-		mon := b.Mon[i%l]
-		// cmd "rbd lock list" serves two purposes:
-		// for fencing, check if lock already held for this host
-		// this edge case happens if host crashes in the middle of acquiring lock and mounting rbd
-		// for defencing, get the locker name, something like "client.1234"
-		cmd, err = b.plugin.execCommand("rbd",
-			append([]string{"lock", "list", b.Image, "--pool", b.Pool, "--id", b.Id, "-m", mon}, secret_opt...))
-		output = string(cmd)
-		glog.Infof("lock list output %q", output)
-		if err != nil {
-			continue
-		}
-
-		if lock {
-			// check if lock is already held for this host by matching lock_id and rbd lock id
-			if strings.Contains(output, lock_id) {
-				// this host already holds the lock, exit
-				glog.V(1).Infof("rbd: lock already held for %s", lock_id)
-				return nil
-			}
-			// clean up orphaned lock if no watcher on the image
-			used, statusErr := util.rbdStatus(&b)
-			if statusErr == nil && !used {
-				re := regexp.MustCompile("client.* " + kubeLockMagic + ".*")
-				locks := re.FindAllStringSubmatch(output, -1)
-				for _, v := range locks {
-					if len(v) > 0 {
-						lockInfo := strings.Split(v[0], " ")
-						if len(lockInfo) > 2 {
-							cmd, err = b.plugin.execCommand("rbd",
-								append([]string{"lock", "remove", b.Image, lockInfo[1], lockInfo[0], "--pool", b.Pool, "--id", b.Id, "-m", mon}, secret_opt...))
-							glog.Infof("remove orphaned locker %s from client %s: err %v, output: %s", lockInfo[1], lockInfo[0], err, string(cmd))
-						}
-					}
-				}
-			}
-
-			// hold a lock: rbd lock add
-			cmd, err = b.plugin.execCommand("rbd",
-				append([]string{"lock", "add", b.Image, lock_id, "--pool", b.Pool, "--id", b.Id, "-m", mon}, secret_opt...))
-		} else {
-			// defencing, find locker name
-			ind := strings.LastIndex(output, lock_id) - 1
-			for i := ind; i >= 0; i-- {
-				if output[i] == '\n' {
-					locker = output[(i + 1):ind]
-					break
-				}
-			}
-			// remove a lock: rbd lock remove
-			cmd, err = b.plugin.execCommand("rbd",
-				append([]string{"lock", "remove", b.Image, lock_id, locker, "--pool", b.Pool, "--id", b.Id, "-m", mon}, secret_opt...))
-		}
-
-		if err == nil {
-			//lock is acquired
-			break
-		}
-	}
-	return err
 }
 
 func (util *RBDUtil) persistRBD(rbd rbdMounter, mnt string) error {
@@ -224,28 +82,16 @@ func (util *RBDUtil) loadRBD(mounter *rbdMounter, mnt string) error {
 	return nil
 }
 
-func (util *RBDUtil) fencing(b rbdMounter) error {
-	// no need to fence readOnly
-	if (&b).GetAttributes().ReadOnly {
-		return nil
-	}
-	return util.rbdLock(b, true)
+type diskMapper interface {
+	MapDisk(disk rbdMounter) (string, error)
+	UnmapDisk(disk rbdMounter, mntPath string) error
 }
 
-func (util *RBDUtil) defencing(c rbdUnmounter) error {
-	// no need to fence readOnly
-	if c.ReadOnly {
-		return nil
-	}
 
-	return util.rbdLock(*c.rbdMounter, false)
 }
 
 func (util *RBDUtil) AttachDisk(b rbdMounter) error {
-	var err error
-	var output []byte
-
-	// create mount point
+	// check the mount point first
 	globalPDPath := b.manager.MakeGlobalPDName(*b.rbd)
 	notMnt, err := b.mounter.IsLikelyNotMountPoint(globalPDPath)
 	// in the first time, the path shouldn't exist and IsLikelyNotMountPoint is expected to get NotExist
@@ -253,65 +99,36 @@ func (util *RBDUtil) AttachDisk(b rbdMounter) error {
 		return fmt.Errorf("rbd: %s failed to check mountpoint", globalPDPath)
 	}
 	if !notMnt {
+		// it looks the specified directory is already a mount point,
+		// so there is no sense in mapping a new rbd device
 		return nil
 	}
 	if err = os.MkdirAll(globalPDPath, 0750); err != nil {
 		return fmt.Errorf("rbd: failed to mkdir %s, error", globalPDPath)
 	}
 
-	devicePath, found := waitForPath(b.Pool, b.Image, 1)
-	if !found {
-		// modprobe
-		_, err = b.plugin.execCommand("modprobe", []string{"rbd"})
-		if err != nil {
-			return fmt.Errorf("rbd: failed to modprobe rbd error:%v", err)
-		}
-
-		// fence off other mappers
-		if err = util.fencing(b); err != nil {
-			return fmt.Errorf("rbd: image %s is locked by other nodes", b.Image)
-		}
-		// rbd lock remove needs ceph and image config
-		// but kubelet doesn't get them from apiserver during teardown
-		// so persit rbd config so upon disk detach, rbd lock can be removed
-		// since rbd json is persisted in the same local directory that is used as rbd mountpoint later,
-		// the json file remains invisible during rbd mount and thus won't be removed accidentally.
-		util.persistRBD(b, globalPDPath)
-
-		// rbd map
-		l := len(b.Mon)
-		// avoid mount storm, pick a host randomly
-		start := rand.Int() % l
-		// iterate all hosts until mount succeeds.
-		for i := start; i < start+l; i++ {
-			mon := b.Mon[i%l]
-			glog.V(1).Infof("rbd: map mon %s", mon)
-			if b.Secret != "" {
-				output, err = b.plugin.execCommand("rbd",
-					[]string{"map", b.Image, "--pool", b.Pool, "--id", b.Id, "-m", mon, "--key=" + b.Secret})
-			} else {
-				output, err = b.plugin.execCommand("rbd",
-					[]string{"map", b.Image, "--pool", b.Pool, "--id", b.Id, "-m", mon, "-k", b.Keyring})
-			}
-			if err == nil {
-				break
-			}
-			glog.V(1).Infof("rbd: map error %v %s", err, string(output))
-		}
-		if err != nil {
-			return fmt.Errorf("rbd: map failed %v %s", err, string(output))
-		}
-		devicePath, found = waitForPath(b.Pool, b.Image, 10)
-		if !found {
-			return errors.New("Could not map image: Timeout after 10s")
-		}
+	// map a block device
+	mapper := &RBDKernel
+	devicePath, err := mapper.MapDisk(b)
+	if err != nil {
+		return fmt.Errorf("rbd: cannot map block device")
 	}
+
+	// rbd lock remove needs ceph and image config while the whole detach requires BackendType
+	// but kubelet doesn't get them from apiserver during teardown
+	// so persit rbd config to let disk detaching as well as rbd unlocking
+	// since rbd json is persisted in the same local directory that is used as rbd mountpoint later,
+	// the json file remains invisible during rbd mount and thus won't be removed accidentally.
+	util.persistRBD(b, globalPDPath)
 
 	// mount it
-	if err = b.mounter.FormatAndMount(devicePath, globalPDPath, b.fsType, nil); err != nil {
-		err = fmt.Errorf("rbd: failed to mount rbd volume %s [%s] to %s, error %v", devicePath, b.fsType, globalPDPath, err)
+	if err := b.mounter.FormatAndMount(devicePath, globalPDPath, b.fsType, nil); err != nil {
+		if err := mapper.UnmapDisk(b, devicePath); err != nil {
+			return fmt.Errorf("rbd detach disk: failed to unmap: %s\nError: %v", devicePath, err)
+		}
+		return fmt.Errorf("rbd: failed to mount rbd volume %s [%s] to %s, error %v", devicePath, b.fsType, globalPDPath, err)
 	}
-	return err
+	return nil
 }
 
 func (util *RBDUtil) DetachDisk(c rbdUnmounter, mntPath string) error {
@@ -324,16 +141,14 @@ func (util *RBDUtil) DetachDisk(c rbdUnmounter, mntPath string) error {
 	}
 	// if device is no longer used, see if can unmap
 	if cnt <= 1 {
-		// rbd unmap
-		_, err = c.plugin.execCommand("rbd", []string{"unmap", device})
-		if err != nil {
-			return fmt.Errorf("rbd: failed to unmap device %s:Error: %v", device, err)
+		// load ceph and image/pool info to remove fencing
+		if err := util.loadRBD(c.rbdMounter, mntPath); err != nil {
+			return fmt.Errorf("rbd detach disk: failed to load the persisted metadata\nError: %v", err)
 		}
 
-		// load ceph and image/pool info to remove fencing
-		if err := util.loadRBD(c.rbdMounter, mntPath); err == nil {
-			// remove rbd lock
-			util.defencing(c)
+		mapper := &RBDKernel
+		if err := mapper.UnmapDisk(*c.rbdMounter, device); err != nil {
+			return fmt.Errorf("rbd detach disk: failed to unmap: %s\nError: %v", device, err)
 		}
 
 		glog.Infof("rbd: successfully unmap device %s", device)
